Fixed error with botUrl
Variable botUrl took the value from botToken before the token was read
from the file. So it was an empty string this time and so there where
the botToken should have been in the botUrl was just - nothing. botUrl
is now taking botToken's value AFTER this was read from the file.